fixed downloading of student code

+++ utils/ps3_grading.py
@@ -18,7 +18,7 @@ def download_with_auth(url, filename):
     import urllib2
     print "downloading %s" % filename
     opener = urllib2.build_opener()
-    opener.addheaders.append(('Cookie', 'tracker=direct; tz=America%2FNew_York; spy_repo=ferristseng%2Fcs4414-ps1; spy_repo_at=2013-09-15T21%3A00%3A59.450Z; __utma=1.1370771186.1377548854.1379274281.1379278858.13; __utmb=1.2.10.1379278858; __utmc=1; __utmz=1.1378316186.7.2.utmcsr=google|utmccn=(organic)|utmcmd=organic|utmctr=(not%20provided); user_session=_gV3EHAa4ZUQoA6vbuaHxjNDOJ2Vn6yy7MCTxI9ZxB-1RP9-; spy_user=cs4414uva; logged_in=yes; dotcom_user=cs4414uva; _gh_sess=BAh7CzoJdXNlcmkDTf1QOhBmaW5nZXJwcmludEkiJTFjNjgzNDcxNjhlYzMxMDk2MThkMTU0YzlhZDVmMmU4BjoGRVQ6Dmxhc3Rfc3Vkb0l1OglUaW1lDTFhHIAuUwSHCToNbmFub19udW1pAoUCOg1uYW5vX2RlbmkGOg1zdWJtaWNybyIHZFA6C29mZnNldGn%2BkJ06D3Nlc3Npb25faWQiJWExZmIzN2RiMzkzM2NmNDljODFkZTA5MzE3ZjNmNzQwOhBfY3NyZl90b2tlbkkiMXBBUFZ3enlQS0ZEUFZLd3JGV3pxTHFReEcvTkJsKzlFSDczWkQyTEN1TGc9BjsHRjoMY29udGV4dEkiBi8GOwdG--f632ad9e4cf92fb137c768dad2d097f609c3d338'))
+    opener.addheaders.append(('Cookie', 'logged_in=yes; dotcom_user=cs4414uva; tz=America%2FNew_York; __utma=1.1596037320.1383061983.1383066523.1383147935.3; __utmb=1.2.10.1383147935; __utmc=1; __utmz=1.1383061983.1.1.utmcsr=(direct)|utmccn=(direct)|utmcmd=(none); user_session=KKSZ4Vc6VKetzRSIVgQxXcrUPPULzUTPzMQwrS-SuUnoEnfw; spy_user=cs4414uva; _gh_sess=BAh7CjoPc2Vzc2lvbl9pZCIlMmU1NzM1NWVjMzc3NzQwOGM0YjcyMzkzZDZjNjkzOTA6DGNvbnRleHRJIgYvBjoGRUY6EF9jc3JmX3Rva2VuSSIxd25lU3dML3ltcjdhV2c4eHdKTlRnMmxvSVVTWkJ2TmtNSlEvdEQvdTRLQT0GOwdGOg1zcHlfcmVwb0kiF1N5ZG5leUgvY3M0NDE0LXBzMQY7B1Q6EHNweV9yZXBvX2F0SXU6CVRpbWUNz2ccgNNT8LYJOg1uYW5vX251bWkC7AE6DW5hbm9fZGVuaQY6DXN1Ym1pY3JvIgdJIDoLb2Zmc2V0af6QnQ%3D%3D--df901dfc656d661bd7e67b592e268d51d56899f7'))
     f = opener.open(url)
     data = f.read()
     with open(filename, "wb") as code:
@@ -34,26 +34,40 @@ def csv_parse(csv_path, tarball_dir):
                 github_username, repo_name, submit_version = tag_match.group(1, 2, 4)
                 zip_src_tarball_url = "https://github.com/%s/%s/archive/%s.zip" % (github_username, repo_name, submit_version)
                 
-                tarball_name = "%s - ps3 - %s.zip" % (partner1_email, submit_version)
-                download_with_auth(zip_src_tarball_url, os.path.join(tarball_dir, tarball_name))
+                try:
+                    tarball_name = "%s - ps3 - %s.zip" % (partner1_email, submit_version)
+                    download_with_auth(zip_src_tarball_url, os.path.join(tarball_dir, tarball_name))
+                except:
+                    try:
+                        tarball_name = "%s - ps3 - %s.zip" % (partner2_email, submit_version)
+                        download_with_auth(zip_src_tarball_url, os.path.join(tarball_dir, tarball_name))
                 
-                if repo_name != REPO_NAME:
-                    print "Student name: %s (%s)" % (partner1_name, partner1_email)
-                    print "[ignore] Repo name error: %s\n" % repo_namenn
+                    except:
+                        try:
+                            tarball_name = "%s - ps3 - %s.zip" % (partner3_email, submit_version)
+                            download_with_auth(zip_src_tarball_url, os.path.join(tarball_dir, tarball_name))
+                        except:
+                            print "Could not download repository for %s, %s, and %s" % (partner1_email, partner2_email, partner3_email)
+                        
+                #if repo_name != REPO_NAME:
+                #    print "Student name: %s (%s)" % (partner1_name, partner1_email)
+                #    print "[ignore] Repo name error: %s\n" % repo_namenn
             else:
                 print "Student name: %s (%s)" % (partner1_name, partner1_email)
                 print "URL error: %s\n" % tag_url
 
 def generate_test_files():
-    os.system("dd if=/dev/urandom of=5K.bin bs=5K count=1")
-    os.system("dd if=/dev/urandom of=5M.bin bs=5M count=1")
-    os.system("dd if=/dev/urandom of=10M.bin bs=10M count=1")
-    os.system("dd if=/dev/urandom of=20M.bin bs=20M count=1")
-    os.system("dd if=/dev/urandom of=40M.bin bs=40M count=1")
-    os.system("dd if=/dev/urandom of=80M.bin bs=80M count=1")
-    os.system("dd if=/dev/urandom of=512M.bin bs=512M count=1")
-    os.system("wget http://www.cs.virginia.edu/~wx4ed/cs4414/ps3/zhtta-test-urls.txt")
-    os.system("tr \"\n\" \"\0\" < zhtta-test-urls.txt > zhtta-test-urls.httperf")
+    if not os.path.isfile("512M.bin"):
+        os.system("dd if=/dev/urandom of=5K.bin bs=5K count=1")
+        os.system("dd if=/dev/urandom of=5M.bin bs=5M count=1")
+        os.system("dd if=/dev/urandom of=10M.bin bs=10M count=1")
+        os.system("dd if=/dev/urandom of=20M.bin bs=20M count=1")
+        os.system("dd if=/dev/urandom of=40M.bin bs=40M count=1")
+        os.system("dd if=/dev/urandom of=80M.bin bs=80M count=1")
+        os.system("dd if=/dev/urandom of=512M.bin bs=512M count=1")
+    if not os.path.isfile("zhtta-test-urls.txt"):
+        os.system("wget http://www.cs.virginia.edu/~wx4ed/cs4414/ps3/zhtta-test-urls.txt")
+        os.system("tr \"\\n\" \"\\0\" < zhtta-test-urls.txt > zhtta-test-urls.httperf")
 #    more small files of different names & sizes - name of file does not imply size
     # For localhost - request a greater number of files
     
@@ -95,8 +109,15 @@ def main():
                 sys.stdout.flush()
                 os.system("../../../cloc.pl --quiet zhtta.rs")
                 
-                os.system("cp " + tarball_dir + "/*.bin ./")
-                os.system("cp " + tarball_dir + "zhtta-test-urls.httperf ./")
+                os.system("cp " + tarball_dir + "/5K.bin ./")
+                os.system("cp " + tarball_dir + "/5M.bin ./")
+                os.system("cp " + tarball_dir + "/10M.bin ./")
+                os.system("cp " + tarball_dir + "/20M.bin ./")
+                os.system("cp " + tarball_dir + "/40M.bin ./")
+                os.system("cp " + tarball_dir + "/80M.bin ./")
+                os.system("cp " + tarball_dir + "/512M.bin ./")
+
+                os.system("cp " + tarball_dir + "/zhtta-test-urls.httperf ./")
                 
                 os.system("make");
                 os.system("./zhtta &");
